requirements.txt
altgraph==0.17
appdirs==1.4.4
aspy.yaml==1.3.0
astroid==2.5.2
astropy==4.1
attrs==20.3.0
autopep8==1.5.6
branca==0.4.2
bresenham==0.2.1
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
chardet==4.0.0
click==6.7
colour==0.1.5
cycler==0.10.0
Cython==0.29.22
distlib==0.3.1
elevation==1.1.2
fasteners==0.16
fbs==0.9.2
filelock==3.0.12
flake8==3.9.0
folium==0.12.1
future==0.18.2
geojson==2.5.0
geojsoncontour==0.4.0
greenlet==1.0.0
haversine==2.3.0
identify==2.2.2
idna==2.10
importlib-metadata==3.10.0
importlib-resources==5.1.2
isort==5.8.0
itur==0.2.1
jedi==0.17.2
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
llvmlite==0.36.0
macholib==1.14
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.5.0
numba==0.53.1
numpy==1.19.5
packaging==20.9
pandas==1.1.5
parso==0.7.1
pathspec==0.8.1
pd==0.0.1
peewee==3.14.4
pefile==2019.4.18
Pillow==8.2.0
pluggy==0.13.1
pre-commit==2.11.1
pycodestyle==2.7.0
pycparser==2.20
pydocstyle==6.0.0
pyerfa==1.7.2
pyflakes==2.3.1
PyInstaller==3.4
pylint==2.7.4
pyparsing==2.4.7
pyproj==3.0.1
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.15.2.0
pyqtlet==0.3.3
PyQtWebEngine==5.15.4
PyQtWebEngine-Qt5==5.15.2
PySide2==5.15.2
python-dateutil==2.8.1
python-dotenv==0.16.0
python-editor==1.0.4
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-srtm==0.4.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.3.17
requests==2.25.1
rope==0.18.0
scipy==1.5.4
shiboken2==5.15.2
sip==6.0.3
six==1.15.0
snowballstemmer==2.1.0
SoundFile==0.10.3.post1
SQLAlchemy==1.4.4
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.4
virtualenv==20.4.3
wrapt==1.12.1
xarray==0.16.2
XlsxWriter==1.3.8
xmltodict==0.12.0
yapf==0.31.0
zipp==3.4.1
